Note about MessageEntity: I assumed your MessageEntity has fields like id, text, createdAt, and user. If your MessageEntity structure is different, please share it so I can adjust the toMessage() function accordingly.

+++ main/java/com/uyscuti/social/circuit/data/model/Dialog.java
@@ -2,9 +2,6 @@ package com.uyscuti.social.circuit.data.model;
 
 import java.util.ArrayList;
 
-/*
- * Created by troy379 on 04.04.17.
- */
 import android.os.Parcel;
 import android.os.Parcelable;
 
@@ -17,13 +14,11 @@ public class Dialog implements IDialog<Message>, Parcelable {
     private String dialogName;
     private ArrayList<User> users;
     private Message lastMessage;
-
     private Boolean isSelected = false;
     private int unreadCount;
 
     public Dialog(String id, String name, String photo,
                   ArrayList<User> users, Message lastMessage, int unreadCount) {
-
         this.id = id;
         this.dialogName = name;
         this.dialogPhoto = photo;
@@ -93,10 +88,9 @@ public class Dialog implements IDialog<Message>, Parcelable {
         users = in.createTypedArrayList(User.CREATOR);
         lastMessage = in.readParcelable(Message.class.getClassLoader());
         unreadCount = in.readInt();
+        // Fixed: Read as byte since writeToParcel writes as byte
         isSelected = in.readByte() != 0;
     }
-
-
 
     public static final Creator<Dialog> CREATOR = new Creator<Dialog>() {
         @Override
@@ -123,6 +117,7 @@ public class Dialog implements IDialog<Message>, Parcelable {
         dest.writeTypedList(users);
         dest.writeParcelable(lastMessage, flags);
         dest.writeInt(unreadCount);
-        dest.writeInt(isSelected ? 1 : 0);
+        // Fixed: Write as byte to match the read operation
+        dest.writeByte((byte) (isSelected ? 1 : 0));
     }
 }
